Refactoring to combine model classes.
