+++ src/main/java/org/opensearch/searchrelevance/plugin/judgments/clickmodel/coec/CoecClickModel.java
@@ -34,11 +34,11 @@ import org.opensearch.search.aggregations.bucket.terms.TermsAggregationBuilder;
 import org.opensearch.search.builder.SearchSourceBuilder;
 import org.opensearch.searchrelevance.plugin.Constants;
 import org.opensearch.searchrelevance.plugin.judgments.clickmodel.ClickModel;
-import org.opensearch.searchrelevance.plugin.judgments.model.ClickthroughRate;
-import org.opensearch.searchrelevance.plugin.judgments.model.Judgment;
-import org.opensearch.searchrelevance.plugin.judgments.model.ubi.event.UbiEvent;
 import org.opensearch.searchrelevance.plugin.judgments.opensearch.OpenSearchHelper;
 import org.opensearch.searchrelevance.plugin.judgments.queryhash.IncrementalUserQueryHash;
+import org.opensearch.searchrelevance.plugin.model.ClickthroughRate;
+import org.opensearch.searchrelevance.plugin.model.Judgment;
+import org.opensearch.searchrelevance.plugin.model.ubi.event.UbiEvent;
 import org.opensearch.searchrelevance.plugin.utils.MathUtils;
 import org.opensearch.transport.client.Client;
 import org.opensearch.transport.client.Requests;
