+++ src/main/java/org/opensearch/searchrelevance/plugin/rest/SearchRelevanceRestHandler.java
@@ -32,9 +32,9 @@ import org.opensearch.search.SearchHit;
 import org.opensearch.searchrelevance.plugin.Constants;
 import org.opensearch.searchrelevance.plugin.judgments.clickmodel.coec.CoecClickModel;
 import org.opensearch.searchrelevance.plugin.judgments.clickmodel.coec.CoecClickModelParameters;
-import org.opensearch.searchrelevance.plugin.judgments.model.SearchConfiguration;
 import org.opensearch.searchrelevance.plugin.judgments.opensearch.OpenSearchHelper;
 import org.opensearch.searchrelevance.plugin.model.GetSearchConfigurationsRequest;
+import org.opensearch.searchrelevance.plugin.model.SearchConfiguration;
 import org.opensearch.searchrelevance.plugin.runners.OpenSearchQuerySetRunner;
 import org.opensearch.searchrelevance.plugin.runners.QuerySetRunResult;
 import org.opensearch.searchrelevance.plugin.samplers.ProbabilityProportionalToSizeQuerySampler;
